+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Disable spellcheck in remote terminal view (Terminus v0.3.32+)
+
 ## [0.4.0] - 2024-08-07
 
 ### Added